Revised hmi-Interface
added setMode() to HMI-Interface

+++ NXT/src/parkingRobot/IPerception.java
@@ -38,7 +38,7 @@ public interface IPerception {
 	 * 
 	 * @return brightness of the groundcolor in percent(%)
 	 */
-	public int getRightLineSensorValue();
+	public double getRightLineSensorValue();
 	
 	/**
 	 * calibrate the two LineSensors
